+++ app/src/main/java/com/gangwonhyuil/gangwonhyuil/ui/community/screen/community/PostItem.kt
@@ -1,6 +1,6 @@
 package com.gangwonhyuil.gangwonhyuil.ui.community.screen.community
 
-import com.gangwonhyuil.gangwonhyuil.ui.community.model.WriterInfo
+import com.gangwonhyuil.gangwonhyuil.ui.community.entity.WriterInfo
 import com.gangwonhyuil.gangwonhyuil.util.base.Eigenvalue
 
 data class PostItem(
