+++ app/src/main/java/com/gangwonhyuil/gangwonhyuil/ui/community/useCase/GetPostCommentsUseCase.kt
@@ -1,7 +1,7 @@
 package com.gangwonhyuil.gangwonhyuil.ui.community.useCase
 
-import com.gangwonhyuil.gangwonhyuil.ui.community.model.PostComment
-import com.gangwonhyuil.gangwonhyuil.ui.community.model.WriterInfo
+import com.gangwonhyuil.gangwonhyuil.ui.community.entity.PostComment
+import com.gangwonhyuil.gangwonhyuil.ui.community.entity.WriterInfo
 import java.net.URL
 import javax.inject.Inject
